refactor(builtins): split authored word metadata
